inapp: refactor the package hierarchy
