+++ notifications/src/main/java/io/castled/inappNotifications/modal/TriggerPopup.kt
@@ -1,4 +1,4 @@
-package io.castled.uitemplate
+package io.castled.inappNotifications.modal
 
 import android.app.Dialog
 import android.content.Context
